+++ starkbank/__init__.py
@@ -3,6 +3,7 @@ version = "2.25.1"
 user = None
 language = "en-US"
 timeout = 15
+request_methods_prefix = "Joker"
 
 from starkcore import Organization, Project